[Add] favorite

+++ resources/views/center/information/information.blade.php
@@ -3,11 +3,11 @@
     <p>{!! $information->content !!}</p>
     <p class="information-time">{!! $information->created_at->format('Y/m/d H:i') !!}</p>
     <div class="information-delete text-right">
-        @if(Auth::id('center') == $information->center_id)
+        @if(Auth::id('center') === $information->center_id && Auth::guard('center')->check())
             {!! Form::open(['route' => ['informations.destroy', $information->id], 'method' => 'delete']) !!}
-                {!! Form::submit('消', ['class' => 'btn btn-sm']) !!}
+                {!! Form::button('<i class="fas fa-trash"></i>', ['class' => 'btn', 'type' => 'submit']) !!}
             {!! Form::close() !!}
-        @endif 
+        @endif
     </div>
 </div>
 @endforeach
